Fix issues at his fuzz runs and optimise runs for CI

+++ src/libs/GovernanceFee.sol
@@ -34,7 +34,7 @@ abstract contract GovernanceFee is Ownable {
     /**
      * @dev Max possible fee.
      */
-    uint256 constant MAX_GOVERNANCE_FEE = 10 ** 18 * 0.05; // 5%
+    uint64 constant MAX_GOVERNANCE_FEE = 10 ** 18 * 0.05; // 5%
 
     /**
      * @notice Current applied governance fee.
